refactor Rank into class

+++ Movepicker.cs
@@ -190,7 +190,7 @@ internal class MovePicker
             moves[i] = new ExtMove(
                 m,
                 Value.PieceValue[(int) Phase.MG][(int)pos.piece_on(Move.to_sq(m))]
-                - new Value(200*Rank.relative_rank(pos.side_to_move(), Move.to_sq(m))));
+                - new Value(200 * (int)Rank.relative_rank(pos.side_to_move(), Move.to_sq(m))));
         }
     }
